backend/python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
pyyaml==6.0.1
torch==2.1.2
numpy==1.24.3
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
httpx==0.25.2
bcrypt==4.1.1
pyjwt==2.8.1
